Use the types defined in the <cstdint> header
More specifically we now use the standard types
  int8_t, int16_t, int32_t, int64_t, uint8_t, uint16_t, uint32_t, uint64_t
The header <cstdint> is new in C++11 (<stdint.h> already existed in C99).

In openmsx.hh we already defined types like 'int32' or 'uint64'. Those are
now removed and replaced with the standard types above.

In many locations where we meant to use an (unsigned) 64 bit type we wrote
'(unsigned) long long'. Those are now replaced with (u)int64_t.

In many locations we need an explicit 32 bit type. Usually we simply used the
type 'int' or 'unsigned' there. In some of those locations we now use
'(u)int32_t'. Though locating/fixing all such locations will take a lot more
time.

Currently when we need a 16 bit type we often use our typedef 'word'. This
makes sense in the context of Z80 emulation, but it doesn't make much sense
outside that context. TODO replace such cases with uint16_t.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13097 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/Clock.hh
@@ -23,8 +23,8 @@ private:
 	// stuff below calculates:
 	//   MASTER_TICKS = MAIN_FREQ / (FREQ_NUM / FREQ_DENOM) + 0.5
 	static_assert(MAIN_FREQ < (1ull << 32), "must fit in 32 bit");
-	static const unsigned long long P = MAIN_FREQ * FREQ_DENOM + (FREQ_NUM / 2);
-	static const unsigned long long MASTER_TICKS = P / FREQ_NUM;
+	static const uint64_t P = MAIN_FREQ * FREQ_DENOM + (FREQ_NUM / 2);
+	static const uint64_t MASTER_TICKS = P / FREQ_NUM;
 	static_assert(MASTER_TICKS < (1ull << 32), "must fit in 32 bit");
 	static const unsigned MASTER_TICKS32 = MASTER_TICKS;
 
@@ -59,7 +59,7 @@ public:
 	  */
 	unsigned getTicksTill(EmuTime::param e) const {
 		assert(e.time >= lastTick.time);
-		unsigned long long result = (e.time - lastTick.time) / MASTER_TICKS;
+		uint64_t result = (e.time - lastTick.time) / MASTER_TICKS;
 #ifdef DEBUG
 		// we don't even want this overhead in devel builds
 		assert(result == unsigned(result));
@@ -81,7 +81,7 @@ public:
 	  * or go past the given time.
 	  * It is not allowed to call this method for a time in the past.
 	  */
-	unsigned long long getTicksTillUp(EmuTime::param e) const {
+	uint64_t getTicksTillUp(EmuTime::param e) const {
 		assert(e.time >= lastTick.time);
 		return (e.time - lastTick.time + MASTER_TICKS - 1) / MASTER_TICKS;
 	}
@@ -89,7 +89,7 @@ public:
 	/** Calculate the time at which this clock will have ticked the given
 	  * number of times (counted from its last tick).
 	  */
-	const EmuTime operator+(uint64 n) const {
+	const EmuTime operator+(uint64_t n) const {
 		return EmuTime(lastTick.time + n * MASTER_TICKS);
 	}
